+++ src/main/java/com/jnape/palatable/lambda/lens/lenses/EitherLens.java
@@ -19,6 +19,8 @@ public final class EitherLens {
      * Convenience static factory method for creating a lens over right values, wrapping them in a {@link Maybe}. When
      * setting, a {@link Maybe#nothing()} value means to leave the {@link Either} unaltered, where as a
      * {@link Maybe#just} value replaces the either with a right over the {@link Maybe}.
+     * <p>
+     * Note that this lens is NOT lawful, since "you get back what you put in" fails for {@link Maybe#nothing()}.
      *
      * @param <L> the left parameter type
      * @param <R> the right parameter type
@@ -32,6 +34,8 @@ public final class EitherLens {
      * Convenience static factory method for creating a lens over left values, wrapping them in a {@link Maybe}. When
      * setting, a {@link Maybe#nothing()} value means to leave the {@link Either} unaltered, where as a
      * {@link Maybe#just} value replaces the either with a left over the {@link Maybe}.
+     * <p>
+     * Note that this lens is NOT lawful, since "you get back what you put in" fails for {@link Maybe#nothing()}.
      *
      * @param <L> the left parameter type
      * @param <R> the right parameter type